Fix style parser import of font-awsome
